+++ services/remotecluster/service.go
@@ -20,11 +20,10 @@ const (
 	RecvChanBuffer                = 50
 	ResultsChanBuffer             = 50
 	ResultQueueDrainTimeoutMillis = 10000
-	MaxConcurrentSends            = 1 // TODO: increase when threading issue fixed
+	MaxConcurrentSends            = 10
 	SendMsgURL                    = "api/v4/remotecluster/msg"
 	SendTimeout                   = time.Minute
 	SendFileTimeout               = time.Minute * 5
-	SendFileMaxQueue              = 100
 	PingURL                       = "api/v4/remotecluster/ping"
 	PingFreq                      = time.Minute
 	PingTimeout                   = time.Second * 15
@@ -57,9 +56,8 @@ type TopicListener func(msg model.RemoteClusterMsg, rc *model.RemoteCluster, res
 // Service provides inter-cluster communication via topic based messages.
 type Service struct {
 	server     ServerIface
-	send       chan sendTask
 	httpClient *http.Client
-	sendFiles  []chan sendFileTask
+	send       []chan interface{}
 
 	// everything below guarded by `mux`
 	mux              sync.RWMutex
@@ -94,14 +92,13 @@ func NewRemoteClusterService(server ServerIface) (*Service, error) {
 
 	service := &Service{
 		server:         server,
-		send:           make(chan sendTask, SendChanBuffer),
 		httpClient:     client,
 		topicListeners: make(map[string]map[string]TopicListener),
 	}
 
-	service.sendFiles = make([]chan sendFileTask, MaxConcurrentSends)
-	for i := range service.sendFiles {
-		service.sendFiles[i] = make(chan sendFileTask, SendFileMaxQueue)
+	service.send = make([]chan interface{}, MaxConcurrentSends)
+	for i := range service.send {
+		service.send[i] = make(chan interface{}, SendChanBuffer)
 	}
 
 	return service, nil
@@ -195,13 +192,8 @@ func (rcs *Service) resume() {
 	}
 
 	// create thread pool for concurrent message sending.
-	for i := 0; i < MaxConcurrentSends; i++ {
-		go rcs.sendLoop(rcs.done)
-	}
-
-	// create thread pool for concurrent file sending.
-	for i := range rcs.sendFiles {
-		go rcs.sendFileLoop(i, rcs.done)
+	for i := range rcs.send {
+		go rcs.sendLoop(i, rcs.done)
 	}
 
 	rcs.server.GetLogger().Debug("Remote Cluster Service active")
